Fix circular buffer on 32-bit machines.
Here, sizeof(gsize) < sizeof(guint64), and the varargs to `g_object_new` become
out of sync.

Testing Done:
Compile and run tests on cross armv7hl.

Reviewed at https://reviews.imfreedom.org/r/182/

+++ libpurple/circularbuffer.c
@@ -340,7 +340,7 @@ purple_circular_buffer_class_init(PurpleCircularBufferClass *klass) {
 PurpleCircularBuffer *
 purple_circular_buffer_new(gsize growsize) {
 	return g_object_new(PURPLE_TYPE_CIRCULAR_BUFFER,
-	                    "grow-size", growsize ? growsize : DEFAULT_BUF_SIZE,
+	                    "grow-size", (guint64)(growsize ? growsize : DEFAULT_BUF_SIZE),
 	                    NULL);
 }
